ProgramClient, event parser, fix ExchangeContext levercoin dep.
