+++ hylo-clients/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod exchange_client;
+pub mod program_client;
 pub mod stability_pool_client;
 pub mod util;
